+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat: object tracking. call `onSave` and `onRemove` only if we previously
   called `onLoad` or `onCreate`.
 * fix: clear our repository when doctrine clears
+* refactor: simplify object repository
 
 ## 2.0.2